Final test completed and ready to be merged
